pyproject.toml
[tool.poetry]
name = "prompt-engineering"
version = "0.1.0"
description = ""
authors = ["Cooper Love <cooperjlove@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "1.51"
python-dotenv = "1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
